Add scoped statistic renaming
Currently a couple of tests for lmerMod, lmerModTest and psych are
breaking because the 'statistic' columns don't line up with the
expected results. There is the function rename_columns() that should
handle this, but changing the renaming dictionary to resolve the
errors causes other tests to fail.

This commit resolves this issue by introducing scoped renaming of the
values in the 'statistic' column: statistic renaming that is specific
to certain tidy_stats-producing functions.